Working on add jobs page

+++ src/pages/JobPage.jsx
@@ -12,6 +12,9 @@ const JobPage = ({deleteJob}) => {
 
 
   const handleDelete = (jobId) =>{
+    const confirm = window.confirm("Are you sure you want to delete this job listing?")
+
+    if(!confirm) return
     deleteJob(jobId)
     return navigate('/jobs')
   }
